refactor(backend): Move AI code to separate module
