+++ backend/api/__init__.py
@@ -3,11 +3,11 @@ from dotenv import load_dotenv
 from fastapi import FastAPI, HTTPException
 from fastapi.responses import StreamingResponse
 from openai import AsyncOpenAI
-from openai.types.chat import ChatCompletionMessageParam
 
 import os
-from typing import Iterable, Literal
+from typing import Literal
 
+from .ai import *
 from .rmp import *
 
 _ = load_dotenv()
@@ -23,77 +23,6 @@ openai = AsyncOpenAI(
 @app.get("/", description="Basic endpoint to verify that the API is running.")
 def status() -> dict[Literal["status"], str]:
     return {"status": "Up and running."}
-
-
-def format_feedback(feedback: dict[Literal["node"], ProfessorRating]) -> str:
-    node = feedback["node"]
-    return f"""<userFeedback>
-<clarityRating>{node['clarityRating']}/5.0</clarityRating>
-<difficultyRating>{node['difficultyRating']}/5.0</difficultyRating>
-<helpfulRating>{node['helpfulRating']}/5.0</helpfulRating>
-<comment>{node['comment']}</comment>
-</userFeedback>"""
-
-
-def create_prof_feedback_prompt(
-    *,
-    id: str,
-    info: ProfessorInfo,
-    course: str,
-    course_display: str,
-) -> list[ChatCompletionMessageParam]:
-    # Used for derived RMP URL, in case users ask for a source
-    numerical_id = base64.b64decode(id).decode()[8:]
-    source_url = f"https://www.ratemyprofessors.com/professor/{numerical_id}"
-
-    prompt_comments = "\n".join(map(format_feedback, info["ratings"]["edges"]))
-
-    return [
-        # pyright: ignore [reportUnknownVariableType]
-        {
-            "role": "system",
-            "content": (
-                "You are an assistant designed to help users get information on their professors at the University of Ottawa. "
-                "The following is information on a particular professor. "
-                "Answer any questions that a user may have about this professor.\n\n"
-                "If a user asks where the data came from, or for this professor's page, you should provide the full `siteUrl` key."
-            ),
-        },
-        {
-            "role": "system",
-            "content": f"""<dataSourceSection>
-<siteName>Rate My Professors</siteName>
-<siteUrl>{source_url}</siteUrl>
-</dataSourceSection>
-<professorDetailsSection>
-<professorName>{info['firstName']} {info['lastName']}</professorName>
-<avgRating>{info['avgRating']}</avgRating>
-<avgDifficulty>{info['avgDifficulty']}</avgDifficulty>
-<wouldTakeAgainPercent>{info['wouldTakeAgainPercent']:.2f}<wouldTakeAgainPercent>
-</professorDetailsSection>
-<courseDetailsSection>
-<courseCode>{course}</courseCode>
-<courseDisplayName>{course_display}</courseDisplayName>
-</courseDetailsSection>
-<userFeedbackSection>
-{prompt_comments}
-</userFeedbackSection>""",
-        },
-    ]
-
-
-async def generate_stream(*, messages: Iterable[ChatCompletionMessageParam]):
-    stream = await openai.chat.completions.create(
-        messages=messages,
-        model=AI_MODEL,
-        temperature=0.5,
-        stream=True,
-        max_tokens=1000,
-    )
-    async for chunk in stream:
-        content = chunk.choices[0].delta.content
-        if content is not None:
-            yield content
 
 
 @app.get(
@@ -118,6 +47,8 @@ async def stream_prof_feedback(id: str, course: str, course_display: str, prompt
     # Stream AI completion to frontend
     return StreamingResponse(
         generate_stream(
+            client=openai,
+            model=AI_MODEL,
             messages=create_prof_feedback_prompt(
                 id=id,
                 info=info,
@@ -126,7 +57,7 @@ async def stream_prof_feedback(id: str, course: str, course_display: str, prompt
             )
             + [
                 {"role": "user", "content": prompt},
-            ]
+            ],
         ),
         media_type="text/plain",
     )
